refactor: Move SlackChatMessenger to use in Decidim apps
