+++ config/initializers/slack.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "slack-ruby-client"
+require_relative "../../lib/slack_chat_messenger"
 
 Slack.configure do |config|
   config.token = ENV["SLACK_API_TOKEN"]
