+++ lib/slack_chat_messenger.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+class SlackChatMessenger
+  def self.notify(channel:, message:)
+    unless channel && message
+      Rails.logger.error "Cannot send messages to slack!"
+      return
+    end
+    client = Slack::Web::Client.new
+    client.chat_postMessage(channel: channel, text: message, as_user: true)
+  end
+end
